Update README.md
I didn’t have a computer at the time so I created the zork file on the browser on my Ipad. I couldn’t get it connected to the GitHub app, but I was able to upload the file on this device. I will be able to acquire a computer soon!